Поддержка новых забаненных символов. Поправил работу с забаненными символами

+++ Doctrina/Checks.cs
@@ -11,26 +11,26 @@ namespace Doctrina
         public static bool Cheks(Form1 form1)
         {
 
-            if (form1.MaxLists > form1.MaxQuestionOnListUint * form1.MaxQuestonRepeatUint * form1.doneBlocks.Count)
-            {
-                form1.OnErrorHappen("Листов больше чем возможных комбинаций");
-                return true;
-            }
-            if (form1.MaxQuestonRepeatUint > form1.doneBlocks.Count)
-            {
-                form1.OnErrorHappen("Количество запросов на страницу слишком большое");
-                return true;
-            }
-            if (form1.MaxQuestionOnListUint > form1.doneBlocks.Count)
-            {
-                form1.OnErrorHappen("Количество вопросов на документ больше, чем самих вопросов");
-                return true;
-            }
-            if (RepeatCheck(form1))
-            {
-                form1.OnErrorHappen("Количество доступных повторов документа слишком мало");
-                return true;
-            }
+            //if (form1.MaxLists > form1.MaxQuestionOnListUint * form1.MaxQuestonRepeatUint * form1.doneBlocks.Count)
+            //{
+            //    form1.OnErrorHappen("Листов больше чем возможных комбинаций");
+            //    return true;
+            //}
+            //if (form1.MaxQuestonRepeatUint > form1.doneBlocks.Count)
+            //{
+            //    form1.OnErrorHappen("Количество запросов на страницу слишком большое");
+            //    return true;
+            //}
+            //if (form1.MaxQuestionOnListUint > form1.doneBlocks.Count)
+            //{
+            //    form1.OnErrorHappen("Количество вопросов на документ больше, чем самих вопросов");
+            //    return true;
+            //}
+            //if (RepeatCheck(form1))
+            //{
+            //    form1.OnErrorHappen("Количество доступных повторов документа слишком мало");
+            //    return true;
+            //}
             //if (IsQuestionEnAllList())
             // {
             //     OnErrorHappen("Количество доступных вопросов на все документы меньше, чем запрошенно");
@@ -70,7 +70,10 @@ namespace Doctrina
         {
             List<DoneBlock> randBlockCopy = CopyDoneBlocks(form1.doneBlocks);
             uint someTimer = 0;
-            bool bannedSymbolMeets = false;
+            bool bannedSymbol1Meets = false;
+            bool bannedSymbol2Meets = false;
+            bool bannedSymbol3Meets = false;
+            bool bannedSymbol4Meets = false;
             for (int listNumber = 0; listNumber < form1.MaxLists;)
             {
                 List<DoneBlock> uniqueQuestion = new List<DoneBlock>();
@@ -79,12 +82,45 @@ namespace Doctrina
                     var randBlock = RandomNumber.Between(0, randBlockCopy.Count - 1);
                     if (!uniqueQuestion.Contains(randBlockCopy[randBlock]))
                     {
+                        if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol1))
+                        {
+                            if (bannedSymbol1Meets) continue;
+                        }
+                        if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol2))
+                        {
+                            if (bannedSymbol2Meets) continue;
+                        }
+                        if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol3))
+                        {
+                            if (bannedSymbol3Meets) continue;
+                        }
+                        if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol4))
+                        {
+                            if (bannedSymbol4Meets) continue;
+                        }
                         if (randBlockCopy[randBlock].AllowPrint(form1.MaxQuestonRepeatUint, form1.DateThenAllowPrint))
                         {
-                            if (randBlockCopy[randBlock].AnswerPath.Contains('!'))
+                            if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol1))
                             {
-                                if (bannedSymbolMeets) continue;
-                                bannedSymbolMeets = true;
+                                bannedSymbol1Meets = true;
+                                uniqueQuestion.Add(randBlockCopy[randBlock]);
+                                ++questions;
+                            }
+                            else if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol2))
+                            {
+                                bannedSymbol2Meets = true;
+                                uniqueQuestion.Add(randBlockCopy[randBlock]);
+                                ++questions;
+                            }
+                            else if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol3))
+                            {
+                                bannedSymbol3Meets = true;
+                                uniqueQuestion.Add(randBlockCopy[randBlock]);
+                                ++questions;
+                            }
+                            else if (randBlockCopy[randBlock].AnswerPath.Contains(BannedSymbols.BannedSymbol4))
+                            {
+                                bannedSymbol4Meets = true;
                                 uniqueQuestion.Add(randBlockCopy[randBlock]);
                                 ++questions;
                             }
@@ -100,6 +136,7 @@ namespace Doctrina
                         form1.OnErrorHappen("Ошибка комбинации возможных вариантов \n\rПроверьте вводимые данные и количество доступных вопросов");
                         ErrorLog.AddNewEntry("Вопросов_на_лист=" + form1.MaxQuestionOnListUint + " | Макс_повторов_вопросов= " +
                                              form1.MaxQuestonRepeatUint + " | Всего_листов= " + form1.MaxLists);
+                        if(File.Exists(form1._chooseFolderPath + @"\" + Form1.fileNameListText))
                         File.Copy(form1._chooseFolderPath + @"\" + Form1.fileNameListText,
                             "ErrorList" + DateTime.Now.Day + "_" + DateTime.Now.Hour + "_" + DateTime.Now.Minute + "_" + DateTime.Now.Second + ".csv");
                         return true;
